fix/PYOKEMON-35-EventRequestData types 분리
